refactor education section
