+++ src/components/EducationSection.css
@@ -1,0 +1,30 @@
+.education-section {
+    padding: 100px 15%;
+}
+
+.education-section h1 {
+    font-size: 36px;
+    font-weight: 900;
+    color: #1f2937;
+    margin: 0;
+}
+
+.education-title {
+    border-bottom: solid #1f2937 4px;
+    margin: 50px 0;
+}
+
+.education-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.education-item-right {
+    width: 200px;
+}
+
+ul {
+    list-style-type: none;
+    padding: 0;
+}
