+++ src/components/buttons/ViewMoreButton.jsx
@@ -3,7 +3,7 @@ import './MediumButtons.css'
 
 function ViewMoreButton({ url }) {
     return (
-        <a href={url}>
+        <a href={url} target="_blank" rel="noreferrer">
             <button className='medium-button view-more'>
                 Learn More
             </button>
